+++ src/minterpy/core/grid.py
@@ -1,16 +1,17 @@
 """
 Module for the generating points (which provides the unisolvent nodes)
 """
-from copy import deepcopy
-from typing import Callable, Optional
+from copy import copy, deepcopy
+from typing import Callable, Optional, Union
 
 import numpy as np
 
 from minterpy.global_settings import ARRAY, INT_DTYPE
-from minterpy.gen_points import gen_chebychev_2nd_order_leja_ordered, gen_points_chebyshev, gen_points_from_values
+from minterpy.gen_points import GENERATING_FUNCTIONS, gen_points_from_values
 
 from minterpy.core.multi_index import MultiIndexSet
 from minterpy.core.tree import MultiIndexTree
+from minterpy.utils.arrays import expand_dim
 from minterpy.utils.verification import (
     check_type,
     check_values,
@@ -19,6 +20,12 @@ from minterpy.utils.verification import (
 )
 
 __all__ = ["Grid"]
+
+# Default generating function
+DEFAULT_FUN = "chebyshev"
+
+# Type alias
+GEN_FUNCTION = Callable[[int, int], np.ndarray]
 
 
 def _gen_unisolvent_nodes(multi_index, generating_points):
@@ -29,59 +36,89 @@ def _gen_unisolvent_nodes(multi_index, generating_points):
     return np.take_along_axis(generating_points, multi_index.exponents, axis=0)
 
 
-DEFAULT_GRID_VAL_GEN_FCT = gen_points_chebyshev
-
-
 # TODO implement comparison operations based on multi index comparison operations and the generating values used
 class Grid:
-    """Datatype for the nodes some polynomial bases are defined on.
+    """A class representing the nodes on which interpolating polynomials live.
 
-    For a definition of these nodes (refered to as unisolvent nodes), see the mathematical introduction.
+    Instances of this class provide the data structure for the unisolvent
+    nodes, i.e., points in a hypercube that uniquely determine
+    a multi-dimensional interpolating polynomial
+    (of a specified multi-index set).
+
+    Parameters
+    ----------
+    multi_index : MultiIndexSet
+        The multi-index set of polynomial exponents that the Grid should
+        support.
+    generating_function : Union[GEN_FUNCTION, str], optional
+        The generating function to construct an array of generating points.
+        One of the built-in generating functions may be selected via
+        a string key.
+        This parameter is optional; if neither this parameter nor
+        ``generating_points`` is specified, the default generating function
+        based on the Leja-ordered Chebyshev-Lobatto nodes is selected.
+    generating_points : :class:`numpy:numpy.ndarray`, optional
+        The generating points of the interpolation grid, a two-dimensional
+        array of floats whose columns are the generating points
+        per spatial dimension. The shape of the array is ``(n + 1, m)``
+        where ``n`` is the maximum polynomial degree in all dimensions
+        and ``m`` is the spatial dimension.
+        This parameter is optional. If not specified, the generating points
+        are created from the default generating function. If specified,
+        then the points must be consistent with any non-``None`` generating
+        function.
 
     Notes
     -----
-    - The multi-index set to construct a :class:`Grid` may not be
+    - The multi-index set to construct a :class:`Grid` instance may not be
       downward-closed. However, building a :class:`.MultiIndexTree` used
       in the transformation between polynomials in the Newton and Lagrange
       bases requires a downward-closed multi-index set.
-
-    .. todo::
-        - insert a small introduction to the purpose of :class:`Grid` here.
-        - refactor the exposed attributes (each needs at least a getter)
-        - naming issues for ``generating_points`` and ``generating_values``
+    - The notion of unisolvent nodes, strictly speaking, relies on the
+      downward-closedness of the multi-index set. If the set is not
+      downward-closed then unisolvency cannot be guaranteed.
     """
-
-    # TODO make all attributes read only!
-
-    _unisolvent_nodes: Optional[ARRAY] = None
-
     def __init__(
         self,
         multi_index: MultiIndexSet,
-        generating_points: Optional[ARRAY] = None,
+        generating_function: Optional[Union[GEN_FUNCTION, str]] = None,
+        generating_points: Optional[np.ndarray] = None,
     ):
+
+        # --- Arguments processing
+
         # Process and assign the multi-index set argument
         self._multi_index = _process_multi_index(multi_index)
 
-        self._generating_function = DEFAULT_GRID_VAL_GEN_FCT
+        # If generating_function and points not specified,
+        # use the default generating function
+        no_gen_function = generating_function is None
+        no_gen_points = generating_points is None
+        if no_gen_function and no_gen_points:
+            generating_function = DEFAULT_FUN
+
+        # Process and assign the generating function
+        self._generating_function = _process_generating_function(
+            generating_function
+        )
 
         # Process and assign the generating points argument
-        self._generating_points = _process_generating_points(
-            generating_points,
-            self._multi_index,
-            self._generating_function,
-        )
+        if no_gen_points:
+            generating_points = self._create_generating_points()
+        self._generating_points = _process_generating_points(generating_points)
 
-        # Get the polynomial degree of the grid
-        self._poly_degree = _get_poly_degree(
-            self._generating_points,
-            self._multi_index,
-        )
+        # --- Post-assignment verifications
 
-        # TODO check if values and points fit together
-        # TODO redundant information.
+        # Verify the polynomial degree of the Grid
+        self._verify_grid_poly_degree()
 
-        self._tree: Optional[MultiIndexTree] = None
+        # Verify generating function and points again if both are specified
+        if not no_gen_function and not no_gen_points:
+            self._verify_matching_gen_function_and_points()
+
+        # --- Lazily-evaluated properties
+        self._unisolvent_nodes = None
+        self._tree = None
 
     # --- Factory methods
     # TODO rename: name is misleading. a generator is something different in python:
@@ -90,8 +127,8 @@ class Grid:
     def from_generator(
         cls,
         multi_index: MultiIndexSet,
-        generating_function: Callable = DEFAULT_GRID_VAL_GEN_FCT,
-    ):
+        generating_function: Callable,
+    ) -> "Grid":
         """
         Constructor from a factory method for the ``generating_values``.
 
@@ -111,6 +148,7 @@ class Grid:
             multi_index.spatial_dimension,
         )
 
+        # TODO: Revise this
         return cls.from_value_set(multi_index, generating_points)
 
     @classmethod
@@ -132,7 +170,10 @@ class Grid:
             generating_values,
             spatial_dimension,
         )
-        return cls(multi_index, generating_points=generating_points)
+        return cls(
+            multi_index,
+            generating_points=generating_points,
+        )
 
     # --- Properties
     @property
@@ -148,6 +189,23 @@ class Grid:
             A multi-index set of polynomial exponents associated with the Grid.
         """
         return self._multi_index
+
+    @property
+    def generating_function(self) -> Optional[Callable[[int, int], np.ndarray]]:
+        """The generating function of the interpolation Grid.
+
+        Returns
+        -------
+        Optional[Callable]
+            The generating function of the interpolation Grid which is used
+            to generate the generating points.
+
+        Notes
+        -----
+        - If the generating function is ``None`` then the Grid may not be
+          manipulated that results in a grid of a higher degree.
+        """
+        return self._generating_function
 
     @property
     def generating_points(self) -> np.ndarray:
@@ -183,7 +241,7 @@ class Grid:
           the polynomial degree of the Grid does not depend on the notion
           of ``lp_degree`` as it is based only on one dimension.
         """
-        return self._poly_degree
+        return len(self.generating_points) - 1
 
     @property
     def unisolvent_nodes(self):
@@ -308,6 +366,46 @@ class Grid:
         multi_indices_new = self.multi_index.add_exponents(exponents)
         return self._new_instance_if_necessary(multi_indices_new)
 
+    def expand_dim(self, new_dimension: int) -> "Grid":
+        """Expand the dimension of the Grid.
+
+        Parameters
+        ----------
+        new_dimension : int
+            The new spatial dimension. It must be larger than or equal
+            to the current dimension of the Grid.
+
+        Returns
+        -------
+        Grid
+            The Grid with expanded dimension.
+
+        Notes
+        -----
+        - If no generating function is available, then the generating points
+          are directly expanded by appending zeros in the last column.
+        """
+        # Expand the dimension of the multi-index set
+        multi_index = self.multi_index.expand_dim(new_dimension)
+
+        if self.generating_function is None:
+            # If generating function is not available,
+            # directly expand the generating points
+            gen_points = expand_dim(self.generating_points, new_dimension)
+
+            # Return a new instance
+            return self.__class__(
+                multi_index,
+                generating_points=gen_points,
+                generating_function=self._generating_function,
+            )
+
+        # Return a new instance
+        return self.__class__(
+            multi_index,
+            generating_function=copy(self._generating_function),
+        )
+
     # --- Special methods: Copies
     # copying
     def __copy__(self):
@@ -323,7 +421,11 @@ class Grid:
         copy.copy
             copy operator form the python standard library.
         """
-        return self.__class__(self.multi_index, self.generating_points)
+        return self.__class__(
+            self._multi_index,
+            generating_function=self._generating_function,
+            generating_points=self._generating_points,
+        )
 
     def __deepcopy__(self, mem):
         """Creates of a deepcopy.
@@ -340,17 +442,20 @@ class Grid:
 
         """
         return self.__class__(
-            deepcopy(self.multi_index),
-            generating_points=deepcopy(self.generating_points),
+            deepcopy(self._multi_index),
+            generating_function=deepcopy(self._generating_function),
+            generating_points=deepcopy(self._generating_points),
         )
 
     # --- Dunder methods: Rich comparison
     def __eq__(self, other: "Grid") -> bool:
         """Compare two instances of Grid for exact equality in value.
 
-        Two instances of :class:`Grid` class is equal in value if and only if
-        both the underlying multi-index sets are equal and
-        the generating points are equal.
+        Two instances of :class:`Grid` class is equal in value if and only if:
+
+        - both the underlying multi-index sets are equal, and
+        - both the generating points are equal, and
+        - both the generating functions are equal.
 
         Parameters
         ----------
@@ -376,7 +481,85 @@ class Grid:
         if not np.array_equal(self.generating_points, other.generating_points):
             return False
 
+        if self.generating_function != other.generating_function:
+            return False
+
         return True
+
+    # --- Private internal methods: not to be called directly from outside
+    def _verify_matching_gen_function_and_points(self):
+        """Verify if the generation function and points match.
+
+        Raises
+        ------
+        ValueError
+            If the generating function generates different points than
+            then ones that are provided.
+        """
+        gen_points = self.generating_function(
+            self.poly_degree,
+            self.spatial_dimension,
+        )
+
+        if not np.array_equal(gen_points, self.generating_points):
+            raise ValueError(
+                "The generating function generates points that are "
+                "inconsistent with the generating points"
+            )
+
+    def _create_generating_points(self) -> np.ndarray:
+        """Construct generating points from the generating function.
+
+        Returns
+        -------
+        :class:`numpy:numpy.ndarray`
+            The generating points of the interpolation grid, a two-dimensional
+            array of floats whose columns are the generating points
+            per spatial dimension. The shape of the array is ``(n + 1, m)``
+            where ``n`` is the maximum polynomial degree in all dimensions
+            and ``m`` is the spatial dimension.
+        """
+        multi_index = self._multi_index
+        poly_degree = np.max(multi_index.exponents)
+        spatial_dimension = multi_index.spatial_dimension
+
+        generating_function = self._generating_function
+
+        return generating_function(poly_degree, spatial_dimension)
+
+    def _verify_grid_poly_degree(self):
+        """Verify if the degree of the Grid is consistent with the multi-index.
+
+        Raises
+        ------
+        ValueError
+            If the degree of the Grid is smaller than the maximum
+            one-dimensional polynomial degree in all dimensions specified
+            in the corresponding multi-index set of polynomial exponents.
+
+        Notes
+        -----
+        - While it perhaps makes sense to store the polynomial degrees of each
+          dimension as the instance property instead of the maximum over all
+          dimensions, this has no use because the generating points have
+          a uniform length in every dimension. For instance, if the maximum
+          polynomial degrees per dimension of a two-dimensional polynomial are
+          ``[5, 3]``, the stored generating points remain ``(5, 2)`` instead of
+          two arrays having lengths of ``5`` and ``3``, respectively.
+        """
+        # The maximum polynomial degree in any dimension of the multi-index
+        # set indicates the largest degree of one-dimensional polynomial
+        # the grid needs to support.
+        poly_degree_multi_index = np.max(self.multi_index.exponents)
+
+        # Both must be consistent; "smaller" multi-index may be contained
+        # in a larger grid, but not the other way around.
+        if poly_degree_multi_index > self.poly_degree:
+            raise ValueError(
+                f"A grid of a polynomial degree {self.poly_degree} "
+                "cannot consist of multi-indices with a maximum polynomial "
+                "degree {poly_degree_multi_index}"
+            )
 
 
 # --- Internal helper functions
@@ -411,21 +594,53 @@ def _process_multi_index(multi_index: MultiIndexSet) -> MultiIndexSet:
     return multi_index
 
 
+def _process_generating_function(
+    generating_function: Optional[Union[GEN_FUNCTION, str]],
+) -> Optional[GEN_FUNCTION]:
+    """Process the generating function given as argument to Grid constructor.
+
+    Parameters
+    ----------
+    generating_function : Union[GEN_FUNCTION, str], optional
+        The generating function to be processed, either ``None``,
+        a dictionary key as string for selecting from the built-in functions,
+        or a callable.
+
+    Returns
+    -------
+    GEN_FUNCTION, optional
+        The generating function as a callable or ``None`` if not specified.
+    """
+    # None is specified
+    if generating_function is None:
+        return generating_function
+
+    # Get the built-in generating function
+    if isinstance(generating_function, str):
+        return GENERATING_FUNCTIONS[generating_function]
+
+    if callable(generating_function):
+        return generating_function
+
+    raise TypeError(
+        f"The generating function {generating_function} is not callable"
+    )
+
+
 def _process_generating_points(
     generating_points: Optional[np.ndarray],
-    multi_index: MultiIndexSet,
-    generating_function: Optional[Callable],
 ) -> np.ndarray:
     """Process the generating points given as an argument to Grid constructor.
+
+    Parameters
+    ----------
+    generating_points : :class:`numpy:numpy.ndarray`, optional
+        The generating points of the grid, a two-dimensional array of floats
+        of shape ``(n, m)`` where ``n`` is the number of points and ``m``
+        is the number of spatial dimension.
     """
-
-    if generating_points is None:
-        poly_degree = multi_index.poly_degree
-        spatial_dimension = multi_index.spatial_dimension
-        generating_points = generating_function(poly_degree, spatial_dimension)
-
     # Check type
-    check_type(generating_points, np.ndarray)
+    check_type(generating_points, np.ndarray, "The generating points")
     # Check array dimension
     check_dimensionality(generating_points, dimensionality=2)
     # No NaN's and inf's
@@ -435,57 +650,3 @@ def _process_generating_points(
 
     # Return processed generating points
     return generating_points
-
-
-def _get_poly_degree(
-    generating_points: np.ndarray,
-    multi_index: MultiIndexSet,
-) -> int:
-    """Get the poly. degree of the grid from the gen. points.
-
-    Parameters
-    ----------
-    generating_points : :class:`numpy:numpy.ndarray`
-        The generating points of the grid, a two-dimensional array of floats
-        of shape ``(n, m)`` where ``n`` is the number of points and ``m``
-        is the number of spatial dimension. Note that ``n - 1`` is
-        the polynomial degree of the grid.
-    multi_index : MultiIndexSet
-        The multi-indices of polynomial exponents the grid needs to support.
-        The polynomial degree of the set must be smaller than or equal to
-        the degree of the grid.
-
-    Returns
-    -------
-    int
-        The polynomial degree of the Grid instance.
-
-    Notes
-    -----
-    - While it perhaps makes sense to store the polynomial degrees of each
-      dimension as the instance property instead of the maximum over all
-      dimensions, this has no use because the generating points have uniform
-      length for all dimension. For instance, if the maximum polynomial degrees
-      of two-dimensional polynomial are ``[5, 3]``, the stored generating
-      points remain ``(5, 2)`` not two arrays having length ``5`` and ``3``,
-      respectively.
-    """
-    # The number of available points (- 1) indicates the highest 1-dimensional
-    # polynomial degree the grid can support (the degree of the grid)
-    poly_degree_gen_points = len(generating_points) - 1
-
-    # ...but the maximum polynomial degree in any dimension of the multi-index
-    # set indicates the largest degree of one-dimensional polynomial
-    # the grid needs to support.
-    poly_degree_multi_index = np.max(multi_index.exponents)
-
-    # Both must be consistent; "smaller" multi-index may be contained
-    # in a larger grid, but not the other way around.
-    if poly_degree_multi_index > poly_degree_gen_points:
-        raise ValueError(
-            f"A grid of polynomial degree {poly_degree_gen_points} "
-            "cannot consist of multi-indices of with maximum polynomial degree"
-            f" {poly_degree_multi_index}"
-        )
-
-    return poly_degree_gen_points